requirements.txt
Flask==3.1.0
openai==1.3.5
gtts==2.5.4
python-dotenv==1.0.1

